Removed SymlinkUtility.cs. Reorganized project structure for the code to live under ProtoPromise_Unity/Assets/Third Party/ with csprojs referencing them. No longer ignore *.meta files.
^^^ This allows using a github link in Unity's package manager ^^^

Removed CSHARP_7_OR_LATER symbol in favor of CSHARP_7_3_OR_NEWER.
Renamed build configurations `Release With Progress` to `Release` and `Release Without Progress` to `Release_NoProgress` (same with Debug).
Added suffix versions for `.0` without progress and `.1` with progress.
Changed UnhandledException.StackTrace to not use base.StackTrace so it won't show methods that should be hidden when Config.DebugCausalityTracer is not set to All.
Fixed FormatStacktrace in Unity when a stack frame is captured that it can't inspect.

TODO: update progress APIs to not fail and not have deprecated errors when progress is disabled, and instead behave the same (just without calling the callback, and have a deprecated warning instead).